fix(jkr): survive Revit 2027 ElementId ctor + make Fix All / Reset fast (#18)
* fix(jkr): widen Revit element ids to long to survive 2027 runtime

new ElementId(intField) bound to ElementId(Int32) — obsolete since the
Revit 2024 64-bit id change, kept in the 2025.* compile ref but removed
from the Revit 2027 runtime. The call compiled, then threw
MissingMethodException on click ("Method not found: ElementId..ctor(Int32)"),
aborting JKR Fix All / Fix / Locate.

Migrate element-id storage int -> long end-to-end so every new ElementId(...)
binds to the runtime-present ElementId(Int64) ctor:

- crash sites: JkrRenameHandler.ApplyRename (the primary rename path),
  JkrFixApplicator preflight, JkrComplianceDashboardPanel.LocateInRevit,
  ComplianceDashboardPanel.SelectElementInRevit
- DTOs: JkrFixAction/FixResult, IssueVm, JkrAuditStore snapshot,
  ComplianceIssueDto/ModelCheckElement/AIRecommendationDto,
  JkrElementData/JkrComplianceCheckV2/JkrFixActionV2
- cascaded signatures/collections: RenameQueue tuple, RenameResult.FailedElementIds
  (HashSet<long>), MakeFixKey, IssueMapper.BlockFix/IsBlocked, ComputeIssueId
- kill the truncation source: JkrBuildingInfoExtractor no longer casts
  elem.Id.Value down to int (also fixes corruption of ids > int.MaxValue)

Existing .jkr_audit.json integer ids deserialize into long unchanged.
Cost subsystem (CostItem/AICostEstimator) left as-is: no ElementId ctor,
cannot hit this crash.

* perf(jkr): stop Fix All from freezing Revit (view guard + fewer regens)

Fix All applied 85 fixes synchronously on the UI thread, each param fix in its
own SubTransaction (a regen per fix), and — with a rendered 3D/perspective view
active — every regen forced an expensive re-render, so Revit sat "Not Responding"
for long stretches before eventually succeeding.

- View guard: JkrRenameHandler switches away from an active View3D to a
  lightweight drafting/plan view for the duration of the batch, restoring it in
  finally (one re-render at the end instead of one per fix).
- Fewer regenerations: only fixes that may bind a fresh shared parameter keep
  per-fix SubTransaction isolation (JkrFixApplicator.FixNeedsIsolation); plain
  writes on already-present params now run in the outer Transaction, so Revit
  regenerates once at commit. The first fix that binds a (param,category) makes
  it present, so later fixes for it take the fast direct-write path — organically
  batching bindings.
- UX: wait cursor + honest progress copy ("Revit may briefly stop responding")
  so a busy Revit doesn't read as a hang.

Preserves the binding-rollback semantics that guard the post-Reset FixableCount
downgrade. Behavior-preserving; only the transaction shape and active view change.

* perf(jkr): chunk Fix All across Idling ticks so Revit never freezes

The one-shot handler applied the whole batch in a single blocking ExternalEvent
on the UI thread, so a large Fix All held the main thread past Windows' ~5s
threshold and the window went "(Not Responding)".

Rework JkrRenameHandler into an Idling-driven state machine: Execute now captures
the work, switches away from a heavy 3D view, subscribes to UIApplication.Idling
and returns immediately. Each Idling tick (OnIdling) applies fixes for a ~120ms
time slice inside one Transaction, commits, reports progress, then yields via
SetRaiseWithoutDelay so Windows keeps pumping messages and the progress bar
animates 0→N. FinishBatch restores the view, unsubscribes, and fires OnCompleted
exactly once — so the panel's post-processing and _appliedFixes tracking (hence
the Reset button) are unchanged.

- Per-item try/catch: a single bad fix is recorded as a failure and the batch
  continues; committed fixes stay reversible via Reset (replaces the old
  all-or-nothing TransactionGroup, which can't span idle ticks anyway).
- Panel: live progress bar via handler.OnProgress; dropped the wait cursor and
  the "may stop responding" copy since the UI now stays responsive.
- Reset reversal also runs chunked (same handler), so it won't freeze either.

Trade-off: Fix All now produces several native undo steps instead of one, which
is irrelevant here because JKR undo is done via the app-level Reset button, not
Ctrl+Z. Single-op limits (e.g. binding a param to a huge category) can still
block briefly since one atomic op can't be sliced.

* perf(jkr): make Reset instant — local restore instead of auto re-scan

Reset used to reverse the fixes AND then run a full backend re-scan (re-extract
the whole model + POST + re-render), which was the slow part now that the reverse
itself is chunked.

A) Drop the auto re-scan. After the reverse the model is already back to its
   pre-fix state, so restore the UI locally: delete the persisted audit and flip
   every issue back to Open, then let the user press Re-scan when they want fresh
   backend verification (RestoreToOpenLocally). No model extraction, no round-trip.
B) Show the live progress bar during the reverse (same chunked handler + OnProgress
   as Fix All) so Reset animates instead of feeling stuck.

Toast now nudges "Click Re-scan to refresh." Reset stays accurate-enough for an
undo; the user re-verifies on demand.

* fix(jkr-ui): severity filter, responsive header, scrollable issue list

Three JKR Compliance panel UI issues:

1. Severity filter — High/Med/Low badges were display-only. Now clickable:
   click a pill to filter the list to that severity, click again to clear
   (PanelVm.ActiveSeverity + ToggleSeverity, one predicate in Refresh()).
   Composes with category chip, status tab, and search (AND). Non-selected
   pills dim to show the active state.

2. Re-scan button clipped at narrow dock widths — the fixed Auto/*/Auto header
   row pushed Re-scan off the right edge. Replaced with a WrapPanel so
   LOD/Discipline/Re-scan wrap to a second line instead of clipping.

3. Issue list not scrollable / panel didn't resize — the panel was a Page hosted
   in a Frame, which let it size to content so the '*' issue-list row got
   unbounded height and the ScrollViewer never engaged. Converted the panel and
   its host to UserControl and host it directly (no Frame); the UserControl now
   stretches to fill the dockable pane, giving the issue list a bounded, scrollable
   height.

Frontend-only; no backend change.

+++ UI/Jkr/ViewModels/IssueVm.cs
@@ -37,7 +37,7 @@ namespace RevitWebAppSync.UI.Jkr.ViewModels
         public string Description { get; set; } = "";
 
         public ElementRef Element { get; set; } = new ElementRef();
-        public int RevitElementId { get; set; }
+        public long RevitElementId { get; set; }
         public string Required { get; set; } = "";
         public string Actual { get; set; } = "";
         public string Example { get; set; } = "";
